refactor(ui): switch to imperative navigation with useNavigate
Replaces `@tanstack/react-router` `Link` components with `useNavigate`
hooks for better control over navigation, especially within `Menu.Item`
and `Button` `onClick` handlers.
Relocates `AssignmentPanel` from `features/courses/assignments` to
`pages/shared/courses/$courseCode/assignments/course-assignments.tsx`
to consolidate related components.
Removes unused `console.log` statements from `content-tree.tsx`.
